Rename module to "cookies" to match functions
