+++ test/lib/verifyFormSubmissionCookieIsSet.test.ts
@@ -1,5 +1,5 @@
+import { verifyFormSubmissionCookieIsSet } from "../../src/lib/cookies";
 import { formSubmissionCookieId } from "../../src/lib/types";
-import { verifyFormSubmissionCookieIsSet } from "../../src/lib/verifyFormSubmissionCookieIsSet";
 
 describe("VerifyFormSubmissionCookieIsSet", () => {
   it("should return true if the form submission cookie is set", () => {
